# src/main/resources/openapi/address-api.yaml
openapi: 3.0.3

info:
  title: Address Finder API
  version: 1.0.0
  description: API contract-first para consulta de CEP, verificação de cobertura e gravação de logs em banco.

servers:
  - url: http://localhost:8080
    description: Local environment

tags:
  - name: Address
    description: Operações relacionadas à consulta de CEP e verificação de cobertura.

paths:
  /address:
    get:
      tags:
        - Address
      summary: Lista endereços consultados
      description: >
        Retorna os logs de endereços consultados e persistidos na base local,
        utilizando paginação. Este endpoint não consulta a API externa; ele apenas
        lista os registros já gravados em banco.
      operationId: getAllAddresses
      parameters:
        - name: page
          in: query
          required: false
          description: Número da página desejada, iniciando em zero.
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
            example: 0
        - name: size
          in: query
          required: false
          description: Quantidade de registros por página.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 10
            example: 10
      responses:
        '200':
          description: Lista paginada de endereços consultados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressPageResponse'
              examples:
                firstPage:
                  summary: Primeira página de registros
                  value:
                    content:
                      - zipCode: "13458870"
                        street: "Estrada do Barreirinho"
                        complement: "até 1750 - lado par"
                        neighborhood: "Residencial Mac Knight"
                        city: "Santa Bárbara D'Oeste"
                        state: "SP"
                        covered: true
                        source: "DATABASE"
                    page: 0
                    size: 10
                    totalElements: 1
                    totalPages: 1
                    last: true
        '400':
          description: Parâmetros de paginação inválidos.

  /address/zip/{zipCode}:
    get:
      tags:
        - Address
      summary: Busca endereço por CEP
      description: >
        Recebe um CEP e busca os dados do endereço. A aplicação consulta primeiro
        a base de dados local. Caso o CEP não exista na base, consulta uma API externa
        ou mockada, verifica se o endereço está dentro da área de cobertura configurada,
        grava o log da consulta em banco e retorna os dados ao consumidor.
      operationId: getAddressByZipCode
      parameters:
        - name: zipCode
          in: path
          required: true
          description: CEP a ser consultado. Aceita formato com ou sem hífen.
          schema:
            type: string
            example: "13458-870"
            minLength: 8
            maxLength: 9
            pattern: '^\d{5}-?\d{3}$'
      responses:
        '200':
          description: Endereço encontrado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressResponse'
              examples:
                addressFromDatabase:
                  summary: Endereço encontrado na base local
                  value:
                    zipCode: "13458870"
                    street: "Estrada do Barreirinho"
                    complement: "até 1750 - lado par"
                    neighborhood: "Residencial Mac Knight"
                    city: "Santa Bárbara D'Oeste"
                    state: "SP"
                    covered: true
                    source: "DATABASE"

                addressFromExternalApi:
                  summary: Endereço encontrado na API externa
                  value:
                    zipCode: "01001000"
                    street: "Praça da Sé"
                    complement: "lado ímpar"
                    neighborhood: "Sé"
                    city: "São Paulo"
                    state: "SP"
                    covered: true
                    source: "EXTERNAL_API"

                notCoveredAddress:
                  summary: Endereço fora da área de cobertura
                  value:
                    zipCode: "20040020"
                    street: "Rua da Quitanda"
                    complement: ""
                    neighborhood: "Centro"
                    city: "Rio de Janeiro"
                    state: "RJ"
                    covered: false
                    source: "EXTERNAL_API"

        '400':
          description: Requisição inválida ou CEP com formato inválido.

        '404':
          description: CEP válido em formato, mas não encontrado na base local nem na API externa.

        '502':
          description: Erro ao consultar a API externa de endereço.

components:
  schemas:
    AddressPageResponse:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
        - last
      properties:
        content:
          type: array
          description: Lista de endereços retornados na página atual.
          items:
            $ref: '#/components/schemas/AddressResponse'

        page:
          type: integer
          format: int32
          description: Número da página atual.
          example: 0

        size:
          type: integer
          format: int32
          description: Quantidade de registros solicitada por página.
          example: 10

        totalElements:
          type: integer
          format: int64
          description: Total de registros existentes na base.
          example: 25

        totalPages:
          type: integer
          format: int32
          description: Total de páginas disponíveis.
          example: 3

        last:
          type: boolean
          description: Indica se a página atual é a última.
          example: false

    AddressResponse:
      type: object
      required:
        - zipCode
        - city
        - state
        - covered
        - source
      properties:
        zipCode:
          type: string
          description: CEP consultado.
          example: "13458870"

        street:
          type: string
          description: Logradouro retornado pela base local ou API externa.
          example: "Estrada do Barreirinho"

        complement:
          type: string
          description: Complemento retornado pela base local ou API externa.
          example: "até 1750 - lado par"

        neighborhood:
          type: string
          description: Bairro retornado pela base local ou API externa.
          example: "Residencial Mac Knight"

        city:
          type: string
          description: Cidade retornada pela base local ou API externa.
          example: "Santa Bárbara D'Oeste"

        state:
          type: string
          description: Estado retornado pela base local ou API externa.
          example: "SP"

        covered:
          type: boolean
          description: Indica se o endereço está dentro da área de cobertura configurada.
          example: true

        source:
          $ref: '#/components/schemas/AddressSource'

    AddressSource:
      type: string
      description: Origem dos dados retornados.
      enum:
        - DATABASE
        - EXTERNAL_API
      example: "DATABASE"